update variable type

+++ src/Main.java
@@ -4,7 +4,7 @@ public class Main {
     public static void main(String[] args) {
     // An application program where the user enters the price of an item and the program computes shipping costs. If the item price is $100 or more, then shipping is free otherwise it is 2% of the price. The program should output the shipping cost and the total price.
 
-        int itemPrice;
+        double itemPrice;
         double shippingCost;
         double totalPrice;
         final double SHIPPING_RATE = 0.02;
@@ -12,8 +12,8 @@ public class Main {
 
         Scanner in = new Scanner(System.in);
         System.out.print(prompt);
-        if (in.hasNextInt()) {
-            itemPrice = in.nextInt();
+        if (in.hasNextDouble()) {
+            itemPrice = in.nextDouble();
             if (itemPrice >= 100) {
                 shippingCost = 0;
                 totalPrice = itemPrice;
@@ -31,9 +31,6 @@ public class Main {
         else {
             System.out.println("Invalid input. Please enter a valid price.");
         }
-
-
-
 
     }
